+++ marshmallow/decorators.py
@@ -5,7 +5,7 @@ These should be imported from the top-level `marshmallow` module.
 Example: ::
 
     from marshmallow import (
-        Schema, pre_load, pre_dump, post_load, validator, ValidationError
+        Schema, pre_load, pre_dump, post_load, validates_schema, ValidationError
     )
 
     class UserSchema(Schema):
@@ -27,7 +27,7 @@ Example: ::
             namespace = 'results' if many else 'result'
             return {namespace: data}
 
-        @validator
+        @validates_schema
         def validate_email(self, data):
             if len(data['email']) < 3:
                 raise ValidationError('Email must be more than 3 characters', 'email')
@@ -41,10 +41,10 @@ PRE_DUMP = 'pre_dump'
 POST_DUMP = 'post_dump'
 PRE_LOAD = 'pre_load'
 POST_LOAD = 'post_load'
-VALIDATOR = 'validator'
+VALIDATES_SCHEMA = 'validates_schema'
 
-def validator(fn=None, raw=False, pass_original=False):
-    """Register a schema-level validator method.
+def validates_schema(fn=None, raw=False, pass_original=False):
+    """Register a schema-level validates_schema method.
 
     By default, receives a single object at a time, regardless of whether ``many=True``
     is passed to the `Schema`. If ``raw=True``, the raw data (which may be a collection)
@@ -53,7 +53,7 @@ def validator(fn=None, raw=False, pass_original=False):
     If ``pass_original=True``, the original data (before unmarshalling) will be passed as
     an additional argument to the method.
     """
-    return tag_processor(VALIDATOR, fn, raw, pass_original=pass_original)
+    return tag_processor(VALIDATES_SCHEMA, fn, raw, pass_original=pass_original)
 
 def pre_dump(fn=None, raw=False):
     """Register a method to invoke before serializing an object. The method